+++ app/Kconfig
@@ -477,6 +477,10 @@ if ZMK_USB_LOGGING
 config ZMK_LOG_LEVEL
 	default 4
 
+# We do this to avoid log loop where logging to USB generates more log messages.
+config USB_CDC_ACM_LOG_LEVEL
+	default 1
+
 config USB_CDC_ACM_RINGBUF_SIZE
 	default 1024
 
@@ -485,6 +489,9 @@ config LOG_BUFFER_SIZE
 
 config LOG_STRDUP_BUF_COUNT
 	default 16
+
+config LOG_PROCESS_THREAD_STARTUP_DELAY_MS
+	default 1000
 
 #ZMK_USB_LOGGING
 endif
