+++ app/include/zmk/usb.h
@@ -23,7 +23,3 @@ enum zmk_usb_conn_state zmk_usb_get_conn_state();
 
 static inline bool zmk_usb_is_powered() { return zmk_usb_get_conn_state() != ZMK_USB_CONN_NONE; }
 static inline bool zmk_usb_is_hid_ready() { return zmk_usb_get_conn_state() == ZMK_USB_CONN_HID; }
-
-#ifdef CONFIG_ZMK_USB
-int zmk_usb_hid_send_report(const uint8_t *report, size_t len);
-#endif /* CONFIG_ZMK_USB */
