+++ app/include/zmk/usb_hid.h
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2020 The ZMK Contributors
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+#pragma once
+
+int zmk_usb_hid_send_report(const uint8_t *report, size_t len);
