+++ app/src/usb.c
@@ -26,41 +26,6 @@ static void raise_usb_status_changed_event(struct k_work *_work) {
 
 K_WORK_DEFINE(usb_status_notifier_work, raise_usb_status_changed_event);
 
-#ifdef CONFIG_ZMK_USB
-
-static const struct device *hid_dev;
-
-static K_SEM_DEFINE(hid_sem, 1, 1);
-
-static void in_ready_cb(const struct device *dev) { k_sem_give(&hid_sem); }
-
-static const struct hid_ops ops = {
-    .int_in_ready = in_ready_cb,
-};
-
-int zmk_usb_hid_send_report(const uint8_t *report, size_t len) {
-    switch (usb_status) {
-    case USB_DC_SUSPEND:
-        return usb_wakeup_request();
-    case USB_DC_ERROR:
-    case USB_DC_RESET:
-    case USB_DC_DISCONNECTED:
-    case USB_DC_UNKNOWN:
-        return -ENODEV;
-    default:
-        k_sem_take(&hid_sem, K_MSEC(30));
-        int err = hid_int_ep_write(hid_dev, report, len, NULL);
-
-        if (err) {
-            k_sem_give(&hid_sem);
-        }
-
-        return err;
-    }
-}
-
-#endif /* CONFIG_ZMK_USB */
-
 enum usb_dc_status_code zmk_usb_get_status() { return usb_status; }
 
 enum zmk_usb_conn_state zmk_usb_get_conn_state() {
@@ -86,19 +51,6 @@ void usb_status_cb(enum usb_dc_status_code status, const uint8_t *params) {
 
 static int zmk_usb_init(const struct device *_arg) {
     int usb_enable_ret;
-
-#ifdef CONFIG_ZMK_USB
-    hid_dev = device_get_binding("HID_0");
-    if (hid_dev == NULL) {
-        LOG_ERR("Unable to locate HID device");
-        return -EINVAL;
-    }
-
-    usb_hid_register_device(hid_dev, zmk_hid_report_desc, sizeof(zmk_hid_report_desc), &ops);
-
-    usb_hid_init(hid_dev);
-
-#endif /* CONFIG_ZMK_USB */
 
     usb_enable_ret = usb_enable(usb_status_cb);
