+++ app/src/usb_hid.c
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2020 The ZMK Contributors
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+#include <device.h>
+#include <init.h>
+
+#include <usb/usb_device.h>
+#include <usb/class/usb_hid.h>
+
+#include <zmk/usb.h>
+#include <zmk/hid.h>
+#include <zmk/keymap.h>
+#include <zmk/event_manager.h>
+
+LOG_MODULE_DECLARE(zmk, CONFIG_ZMK_LOG_LEVEL);
+
+static const struct device *hid_dev;
+
+static K_SEM_DEFINE(hid_sem, 1, 1);
+
+static void in_ready_cb(const struct device *dev) { k_sem_give(&hid_sem); }
+
+static const struct hid_ops ops = {
+    .int_in_ready = in_ready_cb,
+};
+
+int zmk_usb_hid_send_report(const uint8_t *report, size_t len) {
+    switch (zmk_usb_get_status()) {
+    case USB_DC_SUSPEND:
+        return usb_wakeup_request();
+    case USB_DC_ERROR:
+    case USB_DC_RESET:
+    case USB_DC_DISCONNECTED:
+    case USB_DC_UNKNOWN:
+        return -ENODEV;
+    default:
+        k_sem_take(&hid_sem, K_MSEC(30));
+        int err = hid_int_ep_write(hid_dev, report, len, NULL);
+
+        if (err) {
+            k_sem_give(&hid_sem);
+        }
+
+        return err;
+    }
+}
+
+static int zmk_usb_hid_init(const struct device *_arg) {
+    hid_dev = device_get_binding("HID_0");
+    if (hid_dev == NULL) {
+        LOG_ERR("Unable to locate HID device");
+        return -EINVAL;
+    }
+
+    usb_hid_register_device(hid_dev, zmk_hid_report_desc, sizeof(zmk_hid_report_desc), &ops);
+    usb_hid_init(hid_dev);
+
+    return 0;
+}
+
+SYS_INIT(zmk_usb_hid_init, APPLICATION, CONFIG_APPLICATION_INIT_PRIORITY);
